Add check if the DB file exists, pass database parsing testings
Database parsing testings were passed successfully except of test #15 (whole database). The database formatting took incredibly long time and wasn't ended at all. probably because of a very large amount of output symbols. To establish this, parser needs to be changed in some ways. By the way, the file itself was successfully read and parsed, the problem appears when trying to format.